% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogs.R
\docType{data}
\name{dogs}
\alias{dogs}
\title{Auckland Council Animal Management dog data}
\format{
A dataframe with 12 rows and 6 columns:
\describe{
\item{year}{Financial year data was reported}
\item{dogs}{Known dog population}
\item{registered}{Number of known dogs with a current registration}
\item{desexed}{Number of known dogs that have been desexed}
\item{menacing}{Number of known dogs classified as menacing under the Dog Control Act 1996}
\item{dangerous}{Number of known dogs classified as dangerous under the Dog Control Act 1996}
}
}
\source{
\url{https://www.aucklandcouncil.govt.nz/en/plans-policies-bylaws-reports-projects/our-policies/policy-on-dogs.html}
}
\usage{
dogs
}
\description{
Data relating to the Auckland dog population taken from Auckland Council Animal Management annual reports.
}
\keyword{datasets}
