% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforcement.R
\docType{data}
\name{enforcement}
\alias{enforcement}
\title{Auckland Council Animal Management enforcement data}
\format{
A dataframe with 12 rows and 9 columns:
\describe{
\item{year}{Financial year data was reported}
\item{prosecutions}{Total prosecutions or appeals}
\item{infringements}{Total infringements issued}
\item{failure-bylaw}{Infringements issued under s20(5) - Failure to comply with any bylaw authorised by the section}
\item{failure-control}{Infringements issued under s53(1) - Failure to keep dog under control}
\item{failure-disqual}{Infringements issued under s53(1) - Failure to comply with the effects of disqualification}
\item{failure-menacing}{Infringements issued under s33EC(1) - Failure to comply with effects of classification of dog as menacing dog}
\item{failure-register}{Infringements issued under s42 - Failure to register dog}
\item{obstruction-officer}{Infringements issued under s18 - Wilful obstruction of dog control officer or ranger}
}
}
\source{
\url{https://www.aucklandcouncil.govt.nz/en/plans-policies-bylaws-reports-projects/our-policies/policy-on-dogs.html}
}
\usage{
enforcement
}
\description{
Data relating to Auckland Council Animal Management enforcement activities taken from Auckland Council Animal Management annual reports.
}
\keyword{datasets}
