% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locality.R
\docType{data}
\name{locality}
\alias{locality}
\title{Standardised locality mappings for Animal Management data}
\format{
Data frame with columns
\describe{
\item{suburb}{Suburbs contained within the datasets released by Animal Management.}
\item{rfs_no}{Local Boards with jurisdiction over the suburbs found in the Animal Management data.}
\item{rfs_date}{Auckland regions mapped to Local Boards, as defined by Auckland Council online.}
}
}
\source{
Thomas E. Saunders, compiled from data provided by Auckland Council Animal Management as part of an official information request (#8140017948).
}
\usage{
locality
}
\description{
A collection of suburbs compiled from Animal Management data mapped to their corresponding Local Boards and regions. Suburb names were not consistent across separate datasets within the information release, so this table makes it easier to aggregate suburb-level data to higher-level Local Board and regional groupings in a more standardised way.
}
\keyword{datasets}
