% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owners.R
\docType{data}
\name{owners}
\alias{owners}
\title{Auckland Council Animal Management dog ownership data}
\format{
A dataframe with 12 rows and 5 columns:
\describe{
\item{year}{Financial year data was reported}
\item{owners}{Total number of dog owners}
\item{disqualified}{Total number of disqualified dog owners}
\item{probationary}{Total number of probationary dog owners}
\item{rdol}{Total number of dog owners with a Responsible Dog Owners Licence (RDOL)}
}
}
\source{
\url{https://www.aucklandcouncil.govt.nz/en/plans-policies-bylaws-reports-projects/our-policies/policy-on-dogs.html}
}
\usage{
owners
}
\description{
Data relating to dog owners from Auckland Council Animal Management annual reports.
}
\keyword{datasets}
