% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akldogs-package.R
\docType{package}
\name{akldogs-package}
\alias{akldogs}
\alias{akldogs-package}
\title{akldogs: Dog Control Data From Auckland Council}
\description{
Data relating to dog registrations, impounds, and requests for service, compiled by Auckland Council's Animal Management Division and released as part of an official information request.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tesaunders/akldogs}
  \item Report bugs at \url{https://github.com/tesaunders/akldogs/issues}
}

}
\author{
\strong{Maintainer}: Thomas E. Saunders \email{tom.saunders@auckland.ac.nz} (\href{https://orcid.org/0000-0003-1780-0314}{ORCID}) [copyright holder]

}
\keyword{internal}
