% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akldogs-package.R
\docType{package}
\name{akldogs-package}
\alias{akldogs}
\alias{akldogs-package}
\title{akldogs: Auckland Council Animal Management data}
\description{
Auckland Council data relating to dog control, taken from Animal Management annual reports.
}
\author{
\strong{Maintainer}: Tom Saunders \email{tom.saunders@auckland.ac.nz}

}
\keyword{internal}
