% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impounds.R
\docType{data}
\name{impounds}
\alias{impounds}
\title{Auckland Council Animal Management impound data}
\format{
A dataframe with 12 rows and 5 columns:
\describe{
\item{year}{Financial year data was reported}
\item{impounds}{Total number of dogs impounded}
\item{returned}{Total number of impounded dogs returned to or claimed by owner}
\item{euthanised}{Total number of impounded dogs euthanised}
\item{transferred}{Total number of impounded dogs adopted or transferred to a rescue centre}
}
}
\source{
\url{https://www.aucklandcouncil.govt.nz/en/plans-policies-bylaws-reports-projects/our-policies/policy-on-dogs.html}
}
\usage{
impounds
}
\description{
Data relating to impounded dogs taken from Auckland Council Animal Management annual reports.
}
\keyword{datasets}
