% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\docType{data}
\name{registration}
\alias{registration}
\title{Auckland Council Animal Management dog registration data}
\format{
Data frame with columns
\describe{
\item{sheet_name}{Excel sheet from which the row originates in the raw workbook.}
\item{animal_id}{Animal ID number.}
\item{registration_latest_year}{Latest year the animal held a valid registration.}
\item{animal_breed_description}{Primary breed.}
\item{animal_breed2_description}{Secondary breed.}
\item{animal_age}{Age of animal.}
\item{classification}{Current classification under the Dog Control Act.}
\item{suburb_name}{Suburb dog is registered in.}
\item{owner_registration_class}{Whether the owner uses a SuperGold discount or has a RDOL.}
\item{owner_classification_desc}{Whether the owner is subject to a classification under the Dog Control Act.}
\item{animal_registration_group}{Animal registration group.}
\item{animal_gender_description}{Sex of the animal.}
\item{animal_desexed}{Whether or not the animal is desexed.}
\item{address_area}{Region of Auckland dog is registered in.}
\item{dob}{Animal's date of birth.}
}
}
\source{
Auckland Council, official information request (#8140017948).
}
\usage{
registration
}
\description{
Data compiled by Auckland Council's Animal Management division relating to registered dogs in Auckland for financial years 2021-2025.
}
\keyword{datasets}
